+++ packages/midway/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.5](https://github.com/midwayjs/midway/compare/v2.5.4...v2.5.5) (2020-12-15)
+
+
+### Bug Fixes
+
+* aspect wrapper requestContext instance ([#755](https://github.com/midwayjs/midway/issues/755)) ([84193a0](https://github.com/midwayjs/midway/commit/84193a0a50483a0ec8e25b25a17654d4fc77ed1d))
+
+
+
+
+
+## [2.5.4](https://github.com/midwayjs/midway/compare/v2.5.3...v2.5.4) (2020-12-12)
+
+**Note:** Version bump only for package midway
+
+
+
+
+
+## [2.5.3](https://github.com/midwayjs/midway/compare/v2.5.2...v2.5.3) (2020-12-11)
+
+**Note:** Version bump only for package midway
+
+
+
+
+
 ## [2.5.2](https://github.com/midwayjs/midway/compare/v2.5.1...v2.5.2) (2020-12-04)